+++ bin/commands/things-command.js
@@ -19,6 +19,8 @@ var _thingsListTrialOperation = _interopRequireDefault(require("../operations/th
 
 var _thingsRemoveTrialOperation = _interopRequireDefault(require("../operations/things-remove-trial-operation"));
 
+var _thingsAddTrialOperation = _interopRequireDefault(require("../operations/things-add-trial-operation"));
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 class ThingsCommand extends _command.default {
@@ -75,12 +77,16 @@ class ThingsCommand extends _command.default {
             console.log(commandLineUsage(_thingsListTrialOperation.default.usage));
             break;
 
+          case 'add:trial':
+            console.log(commandLineUsage(_thingsAddTrialOperation.default.usage));
+            break;
+
           case 'remove:trial':
             console.log(commandLineUsage(_thingsRemoveTrialOperation.default.usage));
             break;
 
           default:
-            console.log(commandLineUsage([..._thingsListTrialOperation.default.usage, ..._thingsRemoveTrialOperation.default.usage]));
+            console.log(commandLineUsage([..._thingsListTrialOperation.default.usage, ..._thingsAddTrialOperation.default.usage, ..._thingsRemoveTrialOperation.default.usage]));
         }
 
         process.exit(0);
@@ -96,6 +102,8 @@ class ThingsCommand extends _command.default {
 
       if (operation === 'list:trial') {
         await _thingsListTrialOperation.default.run();
+      } else if (operation === 'add:trial') {
+        await _thingsAddTrialOperation.default.run();
       } else if (operation === 'remove:trial') {
         await _thingsRemoveTrialOperation.default.run();
       } else {